# Use backend-aggregated chart data for faster initial render