This change replaces heavy client-side computation in `Home.jsx` with two new backend summary endpoints. Charts now render directly from server-aggregated data, significantly improving first paint and lowering browser CPU/RAM usage.

## What changed
- Added `GET /skills-summary` → returns `{ type, skill, count }[]`
- Added `GET /location-summary` → returns `{ location, count }[]`
- `Home.jsx`
  - Removed `extractSkills()` and manual reduce logic
  - Added `fetchSkillsSummary()` / `fetchLocationSummary()`
  - Location chart now uses `{ name, value }` with `dataKey="name"`, `barKey="value"`
  - Skills charts consume `{ skill, count }` directly

## Before vs After (local dev measurements)

| Metric / Resource | Before | After | Notes |
|---|---:|---:|---|
| First paint / perceived TTI | 6–10s | **~1.5–2s** | Charts visible almost immediately |
| `/skills-summary` | N/A | **~403 ms / 364 B** | Replaces client-side counting |
| `/location-summary` | N/A | **~697 ms / 2.1 KB** | Replaces reduce on `/jobs` |
| `/jobs` | 1.5–4+ s (large payload) | **~1.5 s** | Only used for categories now |
| CPU / RAM | High on client | **Low / stable** | Less JS work |
| Chart render time | Slow (JS-heavy) | **Fast (direct render)** | Recharts/WordCloud |

> Measured with Chrome DevTools on local `npm run dev` (and verified with `npm run build && npm run preview`).

## Risk
Low. New APIs are additive; existing `/jobs` remains unchanged. Heatmap continues to use `/job-locations`.

## Test plan
1. Start backend (`python index.py`) and frontend (`npm run dev`).
2. Verify `/skills-summary` and `/location-summary` in Network tab (<1s each).
3. Ensure charts render within ~2s; switch chart types and category filters.
4. Optional: run `npm run build && npm run preview` and re-measure.

## Rollout
- Merge to feature branch → stage → main.
- Rollback: revert `Home.jsx` if needed; backend summary endpoints are safe to keep.